requirements.txt
# Common dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
agate==1.9.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
daff==1.3.46
dbt-adapters==1.14.3
dbt-common==1.16.0
dbt-core==1.9.3
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.7.4
dbt-snowflake==1.9.2
deepdiff==7.0.1
Deprecated==1.2.18
docx2txt==0.8
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.17.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.14.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
keyring==25.6.0
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.6.0
msgpack==1.1.0
narwhals==1.31.0
networkx==3.2.1
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyJWT==2.10.1
pyOpenSSL==24.3.0
pypdfium2==4.30.1
PySocks==1.7.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
selenium==4.29.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.13.2
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.43.2
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
uvicorn==0.34.2
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
